codellm_data/pyproject.toml
[project]
name = "codellm_data"
version = "0.1.0"
description = "Download and preprocess data required for training"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "dudeperf3ct" }
]
dependencies = [
    "httpx>=0.28.1",
    "polars>=1.35.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "ray[default]>=2.51.1",
    "swh-datasets>=2.0.2",
    "tqdm>=4.67.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "pyrate-limiter>=3.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.5.0",
]

# ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "D",      # pydocstyle
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RET",    # flake8-return
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "PERF",   # perflint
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",     # unused arguments ok in tests
    "PLR2004", # magic values ok in tests
    "D",       # Don't enforce documentation rule
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

# typos configuration
[tool.typos.files]
extend-exclude = [".gitignore", "LICENSE"]

[tool.typos.default]
locale = "en-gb"

# pyrefly configuration
[tool.pyrefly]
python-version = "3.12"
project-excludes = ["**/.[!/.]*", "**/tests", "**/__pycache__"]
project-includes = ["codellm_data"]
search-path = [".", "codellm_data"]
python-interpreter-path = ".venv/bin/python3"
site-package-path = [".venv/lib/python3.12/site-packages"]
